requirements.txt
pandas
numpy
matplotlib
seaborn
mysql-connector-python
sqlalchemy
scikit-learn